Split remote pull request command builder
