+++ Sources/QuillCodeApp/WorkspaceRemoteGitBasicCommandBuilder.swift
@@ -20,23 +20,19 @@ enum WorkspaceRemoteGitBasicCommandBuilder {
             return args.bool("staged") == true ? "git diff --staged" : "git diff"
         case ToolDefinition.gitStage.name:
             let path = try WorkspaceRemoteProjectPath.relativePath(try args.requiredString("path"))
-            return "git add -- \(shellSingleQuoted(path))"
+            return "git add -- \(WorkspaceRemoteShellCommandFormatter.shellSingleQuoted(path))"
         case ToolDefinition.gitRestore.name:
             let path = try WorkspaceRemoteProjectPath.relativePath(try args.requiredString("path"))
             let stagedFlag = args.bool("staged") == true ? " --staged" : ""
-            return "git restore\(stagedFlag) -- \(shellSingleQuoted(path))"
+            return "git restore\(stagedFlag) -- \(WorkspaceRemoteShellCommandFormatter.shellSingleQuoted(path))"
         case ToolDefinition.gitCommit.name:
             let message = try args.requiredString("message").trimmingCharacters(in: .whitespacesAndNewlines)
             guard !message.isEmpty else {
                 throw GitToolError.emptyCommitMessage
             }
-            return "git commit -m \(shellSingleQuoted(message))"
+            return "git commit -m \(WorkspaceRemoteShellCommandFormatter.shellSingleQuoted(message))"
         default:
             throw WorkspaceRemoteGitToolRequestPlannerError.unsupportedTool(call.name)
         }
-    }
-
-    private static func shellSingleQuoted(_ value: String) -> String {
-        WorkspaceTerminalSessionAdapter.shellSingleQuoted(value)
     }
 }
