+++ Sources/QuillCodeApp/WorkspaceRemoteGitPushCommandBuilder.swift
@@ -21,20 +21,20 @@ enum WorkspaceRemoteGitPushCommandBuilder {
         let upstreamArguments = setUpstream ? "-u " : ""
         if let branch = GitInputValidator.trimmedNonEmpty(branch) {
             let branchName = try GitInputValidator.safeName(branch)
-            return "git push \(upstreamArguments)\(shellSingleQuoted(remoteName)) \(shellSingleQuoted(branchName))"
+            return "git push \(upstreamArguments)\(quoted(remoteName)) \(quoted(branchName))"
         }
 
-        let invalidBranchMessage = shellSingleQuoted(String(describing: GitToolError.invalidGitName("$branch")))
+        let invalidBranchMessage = quoted(String(describing: GitToolError.invalidGitName("$branch")))
         let invalidBranchPattern = "*[!\(GitInputValidator.safeNameCharacters)]*"
         return [
             "branch=$(git branch --show-current)",
-            "test -n \"$branch\" || { printf '%s\\n' \(shellSingleQuoted(String(describing: GitToolError.noCurrentBranch))) >&2; exit 1; }",
+            "test -n \"$branch\" || { printf '%s\\n' \(quoted(String(describing: GitToolError.noCurrentBranch))) >&2; exit 1; }",
             "case \"$branch\" in -*|*..*|\(invalidBranchPattern)) printf '%s\\n' \(invalidBranchMessage) >&2; exit 1;; esac",
-            "git push \(upstreamArguments)\(shellSingleQuoted(remoteName)) \"$branch\""
+            "git push \(upstreamArguments)\(quoted(remoteName)) \"$branch\""
         ].joined(separator: " && ")
     }
 
-    private static func shellSingleQuoted(_ value: String) -> String {
-        WorkspaceTerminalSessionAdapter.shellSingleQuoted(value)
+    private static func quoted(_ value: String) -> String {
+        WorkspaceRemoteShellCommandFormatter.shellSingleQuoted(value)
     }
 }
