+++ Tests/QuillCodeParityTests/ParityToolGateTests.swift
@@ -72,6 +72,12 @@ final class ParityToolGateTests: QuillCodeParityTestCase {
         let hunkBuilderText = try Self.appSourceText(named: "WorkspaceRemoteGitHunkCommandBuilder.swift")
         let pushBuilderText = try Self.appSourceText(named: "WorkspaceRemoteGitPushCommandBuilder.swift")
         let pullRequestBuilderText = try Self.appSourceText(named: "WorkspaceRemoteGitHubPullRequestCommandBuilder.swift")
+        let pullRequestBaseBuilderText = try Self.appSourceText(named: "WorkspaceRemoteGitHubPullRequestBaseCommandBuilder.swift")
+        let pullRequestEditBuilderText = try Self.appSourceText(named: "WorkspaceRemoteGitHubPullRequestEditCommandBuilder.swift")
+        let pullRequestReviewBuilderText = try Self.appSourceText(named: "WorkspaceRemoteGitHubPullRequestReviewCommandBuilder.swift")
+        let pullRequestMergeBuilderText = try Self.appSourceText(named: "WorkspaceRemoteGitHubPullRequestMergeCommandBuilder.swift")
+        let pullRequestSupportText = try Self.appSourceText(named: "WorkspaceRemoteGitHubPullRequestCommandSupport.swift")
+        let remoteShellFormatterText = try Self.appSourceText(named: "WorkspaceRemoteShellCommandFormatter.swift")
         let worktreeBuilderText = try Self.appSourceText(named: "WorkspaceRemoteGitWorktreeCommandBuilder.swift")
         let remotePathText = try Self.appSourceText(named: "WorkspaceRemoteProjectPath.swift")
 
@@ -90,7 +96,22 @@ final class ParityToolGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(pushBuilderText.contains("enum WorkspaceRemoteGitPushCommandBuilder"), "Remote git push command construction should live in a focused builder.")
         XCTAssertTrue(pushBuilderText.contains("GitInputValidator.safeName"), "Remote git push command construction should reuse shared git input validation.")
         XCTAssertTrue(pullRequestBuilderText.contains("enum WorkspaceRemoteGitHubPullRequestCommandBuilder"), "Remote GitHub PR command construction should live in a focused builder.")
-        XCTAssertTrue(pullRequestBuilderText.contains("GitHubPullRequestInputValidator.safeSelector"), "Remote GitHub PR command construction should reuse local PR input validation.")
+        XCTAssertTrue(pullRequestBuilderText.contains("WorkspaceRemoteGitHubPullRequestBaseCommandBuilder.command"), "Remote GitHub PR dispatch should delegate creation and read-only commands.")
+        XCTAssertTrue(pullRequestBuilderText.contains("WorkspaceRemoteGitHubPullRequestEditCommandBuilder.command"), "Remote GitHub PR dispatch should delegate edit commands.")
+        XCTAssertTrue(pullRequestBuilderText.contains("WorkspaceRemoteGitHubPullRequestReviewCommandBuilder.command"), "Remote GitHub PR dispatch should delegate review-thread commands.")
+        XCTAssertTrue(pullRequestBuilderText.contains("WorkspaceRemoteGitHubPullRequestMergeCommandBuilder.command"), "Remote GitHub PR dispatch should delegate merge commands.")
+        XCTAssertFalse(pullRequestBuilderText.contains("gh api"), "Remote GitHub PR dispatch should not own GitHub API shell assembly.")
+        XCTAssertFalse(pullRequestBuilderText.contains("GitInputValidator.trimmedNonEmpty"), "Remote GitHub PR dispatch should not own PR argument validation.")
+        XCTAssertFalse(pullRequestBuilderText.contains("args."), "Remote GitHub PR dispatch should not decode individual PR arguments inline.")
+        XCTAssertTrue(pullRequestBaseBuilderText.contains("ToolDefinition.gitPullRequestCreate.name"), "Remote PR base commands should own create/view/checks/diff/checkout dispatch.")
+        XCTAssertTrue(pullRequestBaseBuilderText.contains("WorkspaceRemoteGitHubPullRequestCommandSupport.appendSelector"), "Remote PR base commands should reuse shared selector assembly.")
+        XCTAssertTrue(pullRequestEditBuilderText.contains("GitHubPullRequestInputValidator.safeReviewers"), "Remote PR edit commands should reuse local reviewer validation.")
+        XCTAssertTrue(pullRequestReviewBuilderText.contains("WorkspaceRemoteProjectPath.relativePath"), "Remote PR review comments should reuse remote path validation.")
+        XCTAssertTrue(pullRequestReviewBuilderText.contains("GitHubPullRequestReviewThreadsQuery.graphql"), "Remote PR review thread listing should share the local GraphQL query.")
+        XCTAssertTrue(pullRequestMergeBuilderText.contains("GitHubPullRequestInputValidator.safeMergeFlag"), "Remote PR merge commands should reuse local merge validation.")
+        XCTAssertTrue(pullRequestSupportText.contains("GitHubPullRequestInputValidator.safeSelector"), "Remote PR command support should centralize selector validation.")
+        XCTAssertTrue(pullRequestSupportText.contains("WorkspaceRemoteShellCommandFormatter.command"), "Remote PR command support should centralize shell command quoting.")
+        XCTAssertTrue(remoteShellFormatterText.contains("WorkspaceTerminalSessionAdapter.shellSingleQuoted"), "Remote shell command quoting should live in one shared formatter.")
         XCTAssertTrue(worktreeBuilderText.contains("enum WorkspaceRemoteGitWorktreeCommandBuilder"), "Remote git worktree command construction should live in a focused builder.")
         XCTAssertTrue(worktreeBuilderText.contains("WorkspaceRemoteProjectPath.worktreePath"), "Remote worktree command construction should reuse remote path normalization.")
         XCTAssertTrue(remotePathText.contains("enum WorkspaceRemoteProjectPath"), "Remote path normalization should live in a focused helper.")
